+++ frontend/src/components/widgets/DataFrame/columns/index.ts
@@ -20,7 +20,7 @@ import CheckboxColumn from "./CheckboxColumn"
 import SelectboxColumn from "./SelectboxColumn"
 import ListColumn from "./ListColumn"
 import NumberColumn from "./NumberColumn"
-import { DateTimeColumn, DateColumn, TimeColumn } from "./DateTimeColumn2"
+import DateTimeColumn, { DateColumn, TimeColumn } from "./DateTimeColumn"
 
 import { DateTimeCellRenderer } from "./cells/DateTimeCell"
